requirements.txt
greenlet==3.1.1
paho-mqtt==2.1.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
SQLAlchemy==2.0.36
typing_extensions==4.12.2
